requirements.txt
labelme==5.1.1
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.5.5.64
Pillow==9.4.0
scipy==1.7.3
thop==0.1.1.post2209072238
torch==1.11.0
torchsummary==1.5.1
tqdm==4.64.0
